av-camera-reference-images: move common packages into packagegroup
Add packagegroup "av-camera-reference-image", which contains all
packages that shall be installed in both image variants. This makes
adding packages to all reference images a to lot easier as only the
packagegroup needs to be modified. Also the packagegroup "vimbax-full"
is added. It contains all package for a full Vimba X installation and is
included in the reference image packagegroup.