packagegroup-av-camera-reference-image: add v4l-utils package
Add v4l-utils package to the reference image package group for aarch64
systems. The v4l-utils package can be used to access v4l2 camera using
the cmdline, which is an easy way to get started and perform some tests-
The v4l2viewer package is removed from the packagegroup as it should
only be installed when qt is available and the required append file
already exist.